Break out caching rules to separate file, have env var to toggle caching headers
