+++ packages/app-content-pages/server/server.js
@@ -1,9 +1,9 @@
 const next = require('next')
 const http = require('http')
 
-const DEFAULT_MAX_AGE = process.env.DEFAULT_MAX_AGE || 60 // 1 minute
-const JS_MAX_AGE = process.env.JS_MAX_AGE || 31536000 // 1 year
+const setCacheHeaders = require('./set-cache-headers')
 
+const ENABLE_CACHE = process.env.ENABLE_CACHE || false
 const assetPrefix = process.env.ASSET_PREFIX || ''
 const dev = process.env.NODE_ENV !== 'production'
 const port = parseInt(process.env.PORT, 10) || 3000
@@ -15,7 +15,7 @@ app.prepare().then(() => {
   const server = new http.Server((req, res) => {
     app.setAssetPrefix(assetPrefix)
 
-    if (!dev) {
+    if (ENABLE_CACHE) {
       setCacheHeaders(req, res)
     }
 
@@ -30,18 +30,3 @@ app.prepare().then(() => {
     console.log(`> Ready on http://localhost:${port}`)
   })
 })
-
-function setCacheHeaders (req, res) {
-  let maxAge = DEFAULT_MAX_AGE
-
-  if (isJsRequest(req)) {
-    maxAge = JS_MAX_AGE
-  }
-
-  res.setHeader('Cache-Control', `max-age=${maxAge}`)
-}
-
-function isJsRequest(req) {
-  const regex = /\.js$/i
-  return regex.test(req.path)
-}
